+++ packages/hooks_riverpod/test/framework_test.dart
@@ -81,7 +81,7 @@ void main() {
   testWidgets('AlwaysAliveProviderBase.read(context) inside initState',
       (tester) async {
     final provider = Provider((_) => 42);
-    int result;
+    int? result;
 
     await tester.pumpWidget(
       ProviderScope(
@@ -92,10 +92,6 @@ void main() {
     );
 
     expect(result, 42);
-  });
-
-  test('ProviderScope requires a child', () {
-    expect(() => ProviderScope(child: null), throwsAssertionError);
   });
 
   testWidgets('throws if no ProviderScope found', (tester) async {
@@ -170,7 +166,7 @@ class MockCreateState extends Mock {
 }
 
 class InitState extends StatefulWidget {
-  const InitState({Key key, this.initState}) : super(key: key);
+  const InitState({Key? key, required this.initState}) : super(key: key);
 
   // ignore: diagnostic_describe_all_properties
   final void Function(BuildContext context) initState;