+++ packages/hooks_riverpod/test/provider_test.dart
@@ -33,8 +33,8 @@ void main() {
   });
 
   testWidgets('mounted', (tester) async {
-    ProviderReference providerState;
-    bool mountedOnDispose;
+    late ProviderReference providerState;
+    bool? mountedOnDispose;
     final provider = Provider<int>((ref) {
       providerState = ref;
       ref.onDispose(() => mountedOnDispose = ref.mounted);
@@ -131,7 +131,6 @@ void main() {
   testWidgets('expose value as is', (tester) async {
     var callCount = 0;
     final provider = Provider((ref) {
-      assert(ref != null, '');
       callCount++;
       return 42;
     });
@@ -168,7 +167,6 @@ void main() {
         overrides: [
           provider.overrideWithProvider(
             Provider((ref) {
-              assert(ref != null, '');
               callCount++;
               return 42;
             }),
@@ -186,7 +184,6 @@ void main() {
         overrides: [
           provider.overrideWithProvider(
             Provider((ref) {
-              assert(ref != null, '');
               callCount++;
               throw Error();
             }),